backend/requirements.txt
blinker==1.9.0
boto3==1.35.98
botocore==1.35.98
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
fsspec==2024.12.0
huggingface-hub==0.27.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
pycryptodomex==3.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.10.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
